requirements.txt
asgiref==3.5.0
atomicwrites==1.4.0
attrs==21.4.0
colorama==0.4.4
coverage==6.3
distlib==0.3.4
Django==4.0.1
django-faktura==0.1
django-wkhtmltopdf==3.3.0
filelock==3.4.2
flake8==4.0.1
iniconfig==1.1.1
mccabe==0.6.1
numpy==1.22.1
packaging==21.3
pandas==1.4.0
Pillow==9.0.0
platformdirs==2.4.1
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.7
pytest==6.2.5
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
smartencoding==0.3
sqlparse==0.4.2
toml==0.10.2
tox==3.24.5
tzdata==2021.5
urllib3==1.26.8
virtualenv==20.13.0
